packet/capture: Initial capture buffer support

+++ tests/aat/spec/packet_capture_spec.py
@@ -1,6 +1,10 @@
 from mamba import description, before, after
 from expects import *
 import os
+import socket
+import shutil
+import subprocess
+import tempfile
 
 import client.api
 import client.models
@@ -13,30 +17,114 @@ from common.matcher import (be_valid_packet_capture,
 CONFIG = Config(os.path.join(os.path.dirname(__file__),
                              os.environ.get('MAMBA_CONFIG', 'config.yaml')))
 
+PING = subprocess.check_output('which ping'.split()).decode().rstrip()
 
-def get_first_port_id(api_client):
+def get_nth_port_id(api_client, index):
     ports_api = client.api.PortsApi(api_client)
     ports = ports_api.list_ports()
     expect(ports).not_to(be_empty)
-    return ports[0].id
+    expect(len(ports) > index)
+    return ports[index].id
 
 
-def capture_model(api_client):
+def capture_model(api_client, id = None):
+    if not id:
+        id = get_nth_port_id(api_client, 0)
+
     config = client.models.PacketCaptureConfig(mode='buffer', buffer_size=16*1024*1024)
     capture = client.models.PacketCapture()
-    capture.source_id = get_first_port_id(api_client)
+    capture.source_id = id
     capture.config = config
 
     return capture
 
+def get_interface_address(api_client, interface_id, domain):
+    intf = api_client.get_interface(interface_id)
+    if domain == socket.AF_INET:
+        for protocol in intf.config.protocols:
+            if protocol.ipv4:
+                return protocol.ipv4.static.address
+    elif domain == socket.AF_INET6:
+        for protocol in intf.config.protocols:
+            if protocol.ipv6:
+                return protocol.ipv6.static.address
+    else:
+        raise AttributeError('Unsupported domain')
+
+def ping_command(ping_binary, address, version=socket.AF_INET, count=1):
+    """Generate a ping shell command for Popen"""
+    args = [ping_binary]
+
+    if version == socket.AF_INET6:
+        args.append('-6')
+    else:
+        args.append('-4')
+    # ping interval
+    args.append('-i .2')
+    # ping count
+    args.append('-c %d' % count)
+
+    args.append(address)
+    return ' '.join(args).split()
+
+def do_ping(api_client, ping_binary, src_id, dst_id, domain, count=1):
+    """Perform a ping from src to dst using the specified binary"""
+    shim = Service(CONFIG.shim())
+    dst_ip = get_interface_address(api_client, dst_id, domain)
+
+    with open(os.devnull, 'w') as null:
+        p = subprocess.Popen(ping_command(ping_binary, dst_ip, domain, count),
+                             stdout=null, stderr=null,
+                             env={'LD_PRELOAD': shim.config.path,
+                                  'OP_BINDTODEVICE': src_id})
+        p.wait()
+        expect(p.returncode).to(equal(0))
+
+def get_pcap(api, id, out_file):
+    # Need to use _preload_content to avoid issues with binary data
+    resp = api.get_capture_pcap(id=id, _preload_content=False)
+    with open(out_file, 'wb') as fdst:
+        shutil.copyfileobj(resp, fdst)
+
+def get_pcap_with_wget(id, out_file):
+    base_url = CONFIG.service('dataplane').base_url
+    with open(os.devnull, 'w') as null:
+        p = subprocess.Popen('wget -O %s %s/packet/capture-results/%s/pcap' % (out_file, base_url, id),
+                             stdout=null, stderr=null, shell=True)
+        p.wait()
+        expect(p.returncode).to(equal(0))
+
+def has_scapy():
+    try:
+        import scapy.all
+        return True
+    except Exception as e:
+        print(repr(e))
+        return False
+
+def pcap_icmp_echo_request_count(pcap_file):
+    count = 0
+    icmp_type = scapy.all.ICMP(type='echo-request').type
+    for packet in scapy.all.rdpcap(pcap_file):
+        if 'ICMP' in packet and packet['ICMP'].type == icmp_type:
+            count += 1
+    return count
 
 with description('Packet Capture,', 'packet_capture') as self:
     with description('REST API,'):
 
         with before.all:
-            service = Service(CONFIG.service())
+            service = Service(CONFIG.service('dataplane'))
             self.process = service.start()
             self.api = client.api.PacketCapturesApi(service.client())
+            self.intf_api = client.api.InterfacesApi(self.api.api_client)
+
+            # By default, ping is a privileged process.  We need it unprivileged
+            # to use LD_PRELOAD, so just make a copy as a regular user.
+            self.temp_dir = tempfile.mkdtemp()
+            shutil.copy(PING, self.temp_dir)
+            self.temp_ping = os.path.join(self.temp_dir, os.path.basename(PING))
+            expect(os.path.isfile(self.temp_ping))
 
         with description('invalid HTTP methods,'):
             with description('/packet/captures,'):
@@ -168,6 +256,12 @@ with description('Packet Capture,', 'packet_capture') as self:
                 self.capture = cap
                 result = self.api.start_capture(self.capture.id)
                 expect(result).to(be_valid_packet_capture_result)
+                self.result = result
+
+            with description('get result,'):
+                with it('succeeds'):
+                    result = self.api.get_capture_result(id=self.result.id)
+                    expect(result.id).to(equal(self.result.id))
 
             with description('unfiltered,'):
                 with it('succeeds'):
@@ -190,7 +284,7 @@ with description('Packet Capture,', 'packet_capture') as self:
 
             with description('by source id,'):
                 with it('succeeds'):
-                    results = self.api.list_capture_results(source_id=get_first_port_id(self.api.api_client))
+                    results = self.api.list_capture_results(source_id=get_nth_port_id(self.api.api_client, 0))
                     expect(results).not_to(be_empty)
 
             with description('non-existent source id,'):
@@ -203,29 +297,68 @@ with description('Packet Capture,', 'packet_capture') as self:
         ###
         with description('capture packets,'):
             with description('counters,'):
-                with _it('TODO'):
-                    assert False
+                with it('has packets'):
+                    cap = self.api.create_capture(capture_model(self.api.api_client, 'dataplane-server'))
+                    expect(cap).to(be_valid_packet_capture)
+                    self.capture = cap
+
+                    self.result = self.api.start_capture(self.capture.id)
+                    expect(self.result).to(be_valid_packet_capture_result)
+                    do_ping(self.intf_api, self.temp_ping,
+                            'dataplane-client', 'dataplane-server',
+                            socket.AF_INET, 4)
+                    self.api.stop_capture(self.capture.id)
+                    result = self.api.get_capture_result(id=self.result.id)
+                    expect(result.id).to(equal(self.result.id))
+                    expect(result.packets).to(be_above(4))
+
             with description('pcap,'):
-                with _it('TODO'):
-                    assert False
+                with it('returns pcap'):
+                    cap = self.api.create_capture(capture_model(self.api.api_client, 'dataplane-server'))
+                    expect(cap).to(be_valid_packet_capture)
+                    self.capture = cap
+
+                    self.result = self.api.start_capture(self.capture.id)
+                    expect(self.result).to(be_valid_packet_capture_result)
+                    do_ping(self.intf_api, self.temp_ping,
+                            'dataplane-client', 'dataplane-server',
+                            socket.AF_INET, 4)
+                    self.api.stop_capture(self.capture.id)
+                    result = self.api.get_capture_result(id=self.result.id)
+                    expect(result.id).to(equal(self.result.id))
+                    expect(result.packets).to(be_above(4))
+
+                    out_file = os.path.join(self.temp_dir, 'test.pcap')
+                    get_pcap(self.api, self.result.id, out_file)
+                    expect(os.path.exists(out_file)).to(equal(True))
+                    #get_pcap_with_wget(self.result.id, out_file)
+                    #expect(os.path.exists(out_file)).to(equal(True))
+                    if has_scapy():
+                        expect(pcap_icmp_echo_request_count(out_file)).to(equal(4))
+                    else:
+                        print('\n!!!Install scapy for PCAP validation!!!')
+                    os.remove(out_file)
+
+                    # Make sure HTTP connection is still working
+                    result = self.api.get_capture_result(id=self.result.id)
+                    expect(result.id).to(equal(self.result.id))
+                    expect(result.packets).to(be_above(4))
+
             with description('live,'):
                 with _it('TODO'):
                     assert False
 
         with after.each:
-            try:
+            if hasattr(self, 'api'):
                 for cap in self.api.list_captures():
                     if cap.active:
                         self.api.stop_capture(cap.id)
                 self.api.delete_captures()
-            except AttributeError:
-                pass
             self.capture = None
 
         with after.all:
-            try:
+            if hasattr(self, 'process'):
                 self.process.terminate()
                 self.process.wait()
-            except AttributeError:
-                pass
-
+            if hasattr(self, 'temp_dir') and os.path.isdir(self.temp_dir):
+                shutil.rmtree(self.temp_dir)
